requirements.txt
click==8.0.0
Flask==2.0.0
image==1.5.33
itsdangerous==2.0.0
Jinja2==3.0.0
lxml==4.6.3
MarkupSafe==2.0.0
Pillow==8.2.0
python-docx==0.8.11
pytz==2021.1
qrcode==6.1
six==1.16.0
Werkzeug==2.0.0
